Refactor AdminApp into two components to avoid unnecessary re-renders and loss of scroll position
